Gjør testene letter å forstå
